Looking at this diff, I can see there are a couple of issues with the Git config symlink creation:
1. The comment indicates that `$PWD/config/git` is not the correct path and that the actual location is "two levels" (comment appears cut off)
2. The original code used a relative path `../../config/git` which was replaced with `$PWD/config/git`

The issue is that `$PWD` represents the current working directory when the script is run, but what you actually need is the path relative to where the script is located. Since you're already using `PARENT_FOLDER` to get the script's directory, here's the correct fix:

```bash
# Link Git config if it doesn't exist
[ ! -e ~/.config/git ] && ln -s "$PARENT_FOLDER/git" ~/.config/git
```

This assumes the directory structure is:
```
config/
├── install_git.sh
├── prepare.sh
└── git/           # Git config directory
```

If the git config directory is actually located elsewhere relative to the script, you might need:
```bash
# If git config is in the parent of the config directory
[ ! -e ~/.config/git ] && ln -s "$PARENT_FOLDER/../git" ~/.config/git
```

The key is to use `$PARENT_FOLDER` as the base since it gives you the absolute path to where the script is located, regardless of where the script is run from.